refactor(package): Reservation과 GlobalExceptionHandler를 도메인/예외 패키지로 이동
