+++ src/main/java/roomescape/dto/ReservationRequest.java
@@ -3,7 +3,7 @@ package roomescape.dto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.Reservation;
+import roomescape.domain.Reservation;
 
 public class ReservationRequest {
     private final String name;
