+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -1,12 +1,10 @@
-package roomescape;
+package roomescape.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import roomescape.dto.ExceptionResponse;
-import roomescape.exception.BlankReservationException;
-import roomescape.exception.NotFoundReservationException;
 
 @ControllerAdvice
 public class GlobalExceptionHandler {
