+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.repository;
 
 import java.sql.PreparedStatement;
 import java.time.LocalDate;
@@ -8,6 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.Reservation;
 import roomescape.dto.ReservationRequest;
 import roomescape.exception.NotFoundReservationException;
